[core] Split tests/core/test_io.py to mirror the io package
The 1459-line single file becomes one module per format under
tests/core/io/, with the shared dataset paths and temporary-file cleanup
in `_base.IOTestCase`. All 33 original tests are carried over unchanged;
`test_size_from_slice` is dropped here because it already lives in
test_utils.py alongside the function it covers.

Also unifies the skip idiom. Format-specific classes carry a single
`@unittest.skipUnless(ht.io.supports(...))`, replacing the mix of
`self.skipTest`, `@unittest.skipIf` and -- most importantly -- the
dominant `if ht.io.supports(...): <whole body>` form, which reported a
*pass* for a test that had done nothing. `test_dispatch.py` keeps inline
guards, since those tests deliberately span several formats in one
method.
